@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "jara-utils"
-version = "1.0.0"
+version = "1.1.0"
 description = "A package with basic stuff."
 authors = ["Serban Senciuc <senciuc.serban@gmail.com>"]
 readme = "README_pypi.rst"
@@ -16,13 +16,10 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9.16"
-pytz = "^2022.7.1"
-types-pytz = "^2022.7.1.0"
+python = "^3.11.4"
+pytz = "*"
+types-pytz = "*"
 
-
-[tool.poetry.group.test]
-optional = true
 
 [tool.poetry.group.test.dependencies]
 faker = "^16.4.0"
@@ -30,33 +27,80 @@ pytest = "^7.2.1"
 pytest-cov = "^4.0.0"
 pytest-asyncio = "^0.20.3"
 
-[tool.poetry.group.dev]
-optional = true
-
 [tool.poetry.group.dev.dependencies]
 pytest-sugar = "^0.9.6"
 pytest-deadfixtures = "^2.2.1"
 ipython = "^8.8.0"
 ipdb = "^0.13.11"
 
-[tool.poetry.group.lint]
-optional = true
-
 [tool.poetry.group.lint.dependencies]
 mypy = "^0.991"
 bandit = "^1.7.4"
-flake8 = "^6.0.0"
-flake8-quotes = "^3.3.2"
-flake8-pytest-style = "^1.6.0"
-flake8-comprehensions = "^3.10.1"
-flake8-multiline-containers = "^0.0.19"
-flake8-builtins = "^2.1.0"
-flake8-print = "^5.0.0"
-flake8-debugger = "^4.1.2"
-flake8-simplify = "^0.19.3"
-flake8-annotations = "^2.9.1"
-flake8-commas = "^2.1.0"
+ruff = "^0.0.284"
+black = "^23.7.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+# -------------------- ruff --------------------
+[tool.ruff]
+src = ["src", "test"]
+format = "grouped"
+select = ["ALL"]
+ignore = ["D", "S", "DTZ", "FBT", "ANN101", "ANN102", "ANN401", "ANN002", "ANN003"]
+
+# Allow autofix for all enabled rules (when `--fix`) is provided.
+fixable = ["ALL"]
+unfixable = []
+
+per-file-ignores = {}
+line-length = 120
+target-version = "py311"
+
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+[tool.ruff.flake8-quotes]
+docstring-quotes = "double"
+inline-quotes = "single"
+multiline-quotes = "single"
+
+[tool.ruff.isort]
+combine-as-imports = true
+known-first-party = ["src"]
+
+# -------------------- black --------------------
+[tool.black]
+line-length = 120
+skip-string-normalization = true
+target-version = ["py311"]
+
+# -------------------- pytest --------------------
+#[tool.pytest]
+
+[tool.pytest.ini_options]
+addopts = "-v --cov=jara_utils --cov-report=term --no-cov-on-fail"
+testpaths = ["tests"]
+asyncio_mode = "auto"
+
+# -------------------- pytest-cov --------------------
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "def __repr__",
+    "raise AssertionError",
+    "raise NotImplementedError",
+    "pass",
+    "if 0:",
+    "if __name__ == .__main__.:",
+    "nocov",
+    "if TYPE_CHECKING:",
+    "@abstractmethod",
+]
+fail_under = 95
+show_missing = true
+
+[tool.coverage.run]
+branch = true
+omit = ["tests/*"]
