@@ -1,34 +1,33 @@
 import os
+from typing import Any
 
-from typing import Optional, List, Any
-
-from jara_utils.exceptions import EnvironmentVariableNotFound
+from jara_utils.exceptions import EnvironmentVariableNotFoundError
 from jara_utils.normalization.boolean import str_2_bool
 from jara_utils.normalization.integer import parse_int
 
 
-def get_str(name: str, invalid: Optional[str] = None, required: bool = False) -> Optional[str]:
+def get_str(name: str, invalid: str | None = None, required: bool = False) -> str | None:
     if name not in os.environ and required:
-        raise EnvironmentVariableNotFound(name)
+        raise EnvironmentVariableNotFoundError(name)
     return str(os.environ.get(name, '')) or invalid
 
 
-def get_int(name: str, invalid: Optional[int] = None, required: bool = False) -> Optional[int]:
+def get_int(name: str, invalid: int | None = None, required: bool = False) -> int | None:
     if name not in os.environ and required:
-        raise EnvironmentVariableNotFound(name)
+        raise EnvironmentVariableNotFoundError(name)
     return parse_int(os.environ.get(name)) or invalid
 
 
-def get_bool(name: str, invalid: Optional[bool] = None, required: bool = False) -> Optional[bool]:
+def get_bool(name: str, invalid: bool | None = None, required: bool = False) -> bool | None:
     if name not in os.environ and required:
-        raise EnvironmentVariableNotFound(name)
+        raise EnvironmentVariableNotFoundError(name)
     value = str_2_bool(os.environ.get(name))
     return value if value is not None else invalid
 
 
-def get_list(name: str, separator: str = ',', invalid: Optional[List[Any]] = None, required: bool = False) -> List[str]:
+def get_list(name: str, separator: str = ',', invalid: list[Any] | None = None, required: bool = False) -> list[str]:
     if name not in os.environ and required:
-        raise EnvironmentVariableNotFound(name)
+        raise EnvironmentVariableNotFoundError(name)
 
     if invalid is None:
         invalid = []
