@@ -1,15 +1,15 @@
-from datetime import datetime, date
-from typing import Optional, Any
+from datetime import date, datetime
+from typing import Any
 
 
-def parse_int(value: Any) -> Optional[int]:
+def parse_int(value: Any) -> int | None:
     try:
         return int(value)
     except (ValueError, TypeError):
         return None
 
 
-def age_from(birthday: Optional[datetime]) -> Optional[int]:
+def age_from(birthday: datetime | None) -> int | None:
     if not birthday or birthday == datetime.min:
         return None
     today = date.today()
