@@ -1,16 +1,14 @@
+from collections.abc import AsyncGenerator, AsyncIterable
 from io import StringIO
-from typing import AsyncIterable, List
 
 import pytest
 from faker import Faker
-
 from jara_utils.cli_output import CLIOutput
 
 
 @pytest.fixture(scope='session')
 def faker() -> Faker:
-    f = Faker()
-    return f
+    return Faker()
 
 
 @pytest.fixture()
@@ -29,13 +27,14 @@ def cli_output(stream_buffer: StringIO, error_buffer: StringIO) -> CLIOutput:
 
 
 @pytest.fixture()
-def random_numbers(faker: Faker) -> List[int]:
+def random_numbers(faker: Faker) -> list[int]:
     return faker.pylist(faker.pyint(10, 80), False, [int])
 
 
 @pytest.fixture()
-def random_number_async_iterable(faker: Faker, random_numbers: List[int]) -> AsyncIterable:
-    async def generator():
+def random_number_async_iterable(random_numbers: list[int]) -> AsyncIterable:
+    async def generator() -> AsyncGenerator[int, None]:
         for number in random_numbers:
             yield number
+
     return generator()
