@@ -1,10 +1,8 @@
 from datetime import datetime
-from typing import Union, List, Optional
 
 import pytest
 from faker import Faker
-
-from jara_utils.normalization import parse_int, age_from
+from jara_utils.normalization import age_from, parse_int
 
 
 @pytest.mark.parametrize(
@@ -16,12 +14,13 @@ from jara_utils.normalization import parse_int, age_from
         ('string', None),
         ([], None),
         (datetime.now(), None),
-    ])
-def test_parse_int(value: Union[int, str, float, List, datetime], result: Optional[int]):
+    ],
+)
+def test_parse_int(value: str | float | list | datetime, result: int | None) -> None:
     assert parse_int(value) == result
 
 
-def test_age_from(faker: Faker):
+def test_age_from(faker: Faker) -> None:
     age = faker.pyint(min_value=18, max_value=60)
     assert age_from(None) is None
     assert age_from(datetime.min) is None
