@@ -1,11 +1,11 @@
 class TextColor:
-    ENDC = '\033[0m'  # default format
+    ENDC = DEFAULT = '\033[0m'
 
-    CYAN = '\033[0;36m'  # debug
-    OKBLUE = '\033[94m'  # info
-    OKGREEN = '\033[92m'  # success
-    YELLOW = '\033[93m'  # warning
-    RED = '\033[91m'  # error/fail
+    CYAN = DEBUG = '\033[0;36m'
+    OKBLUE = INFO = '\033[94m'
+    OKGREEN = SUCCESS = '\033[92m'
+    YELLOW = WARNING = '\033[93m'
+    RED = ERROR = FAIL = '\033[91m'
 
     BOLD = '\033[1m'
     UNDERLINE = '\033[4m'
